Using uint64_t for all rolling values

+++ cdedup/intset.h
@@ -7,7 +7,7 @@
 typedef struct _Bucket {
   uint32_t slot_count;
   uint32_t used_slots;
-  uint32_t* slots; // An array of size "slot_count" where the first "used_slots" contains valid data.
+  uint64_t* slots; // An array of size "slot_count" where the first "used_slots" contains valid data.
 } Bucket;
 
 typedef struct _IntSet {
@@ -17,8 +17,8 @@ typedef struct _IntSet {
 } IntSet;
 
 IntSet* create_intset(uint32_t bucket_count);
-void add_intset(IntSet* intset, uint32_t int_to_add);
-int contains_intset(IntSet* intset, uint32_t int_to_find);
+void add_intset(IntSet* intset, uint64_t int_to_add);
+int contains_intset(IntSet* intset, uint64_t int_to_find);
 void destroy_intset(IntSet* intset);
 
 #endif
